requirements.txt
fastapi
uvicorn
python-multipart
pymongo
boto3
python-dotenv
minio